feat: Add logout button to index.html
The code changes in this commit add a logout button to the index.html file. This button allows users to log out of the application by removing the token from local storage and redirecting them to the login page. This feature enhances the usability of the application by providing a convenient way for users to log out.